Extract synced inputs into separate file
