+++ indico/web/client/js/react/components/syncedInputs.module.scss
